+++ components/AssignmentCard.tsx
@@ -10,13 +10,13 @@ import UserContext from '../components/UserContext'
 export default function AssignmentCard({ assignment }: { assignment: Assignment }) {
     const router = useRouter();
     const [assignmentState, setAssignment] = useState(assignment);
-    console.log(assignment);
+    const subject = router.query.subjectId;
 
     
 
     return (
         <div className={styles.assignmentcontainer} >
-            <div className={styles.assignmentcard} onClick={() => router.push(`assignments/edit?assignmentId=${assignment.id}`)}>
+            <div className={styles.assignmentcard} onClick={() => router.push(`/assignments/${subject}/${assignment.id}`)}>
                 <div className={styles.assignmenthead}>
                     <p>{assignmentState.subject.name}</p>
                 </div>
