+++ components/SubjectItem.tsx
@@ -1,10 +1,10 @@
 import { useRouter } from 'next/router';
 import styles from '../styles/SubjectItem.module.scss'
 import Image from "next/image";
-export default function SubjectItem({subject = "DBI",picture}: {subject: string, picture?: boolean}) {
+export default function SubjectItem({subject = "DBI",picture,link}: {subject: string, picture?: boolean, link?: string}) {
     const router = useRouter();
     function openCollection() {
-        router.push(`/summaries/collection?subject=${subject}`);
+        router.push(`${link}/${subject}`);
     }
     return (
         <div className={styles.container} onClick={openCollection}>
