+++ styles/Assignment.module.scss
@@ -1079,6 +1079,45 @@
   justify-content: center;
   align-items: center;
 
+  
+  &>div:first-child{
+    width: 80%;
+    display: inline;
+    margin: 1em 0;
+    vertical-align: middle;
+    position: relative;
+
+    &>*{
+        display: inline;
+        text-align: center;
+        vertical-align: middle;
+    }
+
+    &>img{
+        float: left;
+        width: 3em;
+        height:3em;
+        transform: rotate(180deg);
+        transition: all .2s linear;
+        user-select: none;
+
+        &:hover{
+            filter: brightness(.7);
+            cursor: pointer;
+        }
+    }
+
+    &>h1{
+        text-align: center;
+        margin: auto 0;
+        // center the text
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+    }
+}
+
   &>button{
     width: 60%;
     display: flex;
@@ -1090,6 +1129,7 @@
     margin: .3em 0;
     padding: .5em .8em;
     cursor: pointer;
+
 
     &:hover{
         filter: brightness(1.2);
